feat: bind post with user

+++ src/main/java/com/example/blogplatform/post/PostService.java
@@ -1,13 +1,14 @@
 package com.example.blogplatform.post;
 
-import com.example.blogplatform.exception.errors.BadRequestException;
 import com.example.blogplatform.exception.errors.NotFoundException;
 import com.example.blogplatform.fs.FileSystemService;
 import com.example.blogplatform.post.dto.PostCreateDto;
+import com.example.blogplatform.user.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class PostService {
@@ -33,22 +34,37 @@ public class PostService {
     return postRepository.findByTitleLikeIgnoreCase("%" + title + "%");
   }
 
-  public Post createOne(PostCreateDto dto) {
+  public Post createOne(PostCreateDto dto, User current) {
     String imagePath = fileSystemService.uploadBase64ToSftp(dto.getImage().getBase64());
     Post create = Post.builder()
             .body(dto.getBody())
             .title(dto.getTitle())
+            .user(User.builder().id(current.getId()).build())
             .image(imagePath)
             .build();
 
     return postRepository.save(create);
   }
 
-  public void delete(Long id) {
+  public void delete(Long id, User current) {
+    Optional<Post> exist = postRepository.findByIdAndUserId(id, current.getId());
+    if(exist.isEmpty()){
+      throw new NotFoundException("Post not found");
+    }
+
     postRepository.deleteById(id);
   }
 
-  public Post update(Post post) {
+  public Post update(Post post, User current) {
+    Optional<Post> exist = postRepository.findByIdAndUserId(post.getId(), current.getId());
+    if(exist.isEmpty()){
+      throw new NotFoundException("Post not found");
+    }
+
     return postRepository.save(post);
+  }
+
+  public List<Post> findByUser(Long userId) {
+    return postRepository.findByUserId(userId);
   }
 }
